setup(tailwind): created separate tailwind package and used it in all dependencies
setup(tailwind): bug fix in postcss

setup(tailwind): fix pnpm lock file

removed unused dependencies
