+++ apps/learner/postcss.config.js
@@ -1,1 +1,1 @@
-module.exports = require('@repo/ui/postcss.config');
+module.exports = require('@repo/tailwind-config/postcss-config');
