+++ packages/tailwind-config/tailwind.config.ts
@@ -1,0 +1,17 @@
+import type { Config } from "tailwindcss";
+
+const config = {
+  darkMode: ["class"],
+  content: [
+    "./pages/**/*.{ts,tsx}",
+    "./components/**/*.{ts,tsx}",
+    "./app/**/*.{ts,tsx}",
+    "./src/**/*.{ts,tsx}",
+    "../../packages/ui/src/**/*.{ts,tsx}",
+  ],
+  prefix: "",
+  theme: {},
+  plugins: [],
+} satisfies Config;
+
+export default config;
