+++ packages/ui/postcss.config.js
@@ -1,6 +1,1 @@
-module.exports = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
+module.exports = require('@repo/tailwind-config/postcss-config');
