+++ packages/ui/tailwind.config.ts
@@ -1,17 +1,2 @@
-import type { Config } from "tailwindcss";
+export * from "@repo/tailwind-config/tw-config";
 
-const config = {
-  darkMode: ["class"],
-  content: [
-    "./pages/**/*.{ts,tsx}",
-    "./components/**/*.{ts,tsx}",
-    "./app/**/*.{ts,tsx}",
-    "./src/**/*.{ts,tsx}",
-    "../../packages/ui/src/**/*.{ts,tsx}",
-  ],
-  prefix: "",
-  theme: {},
-  plugins: [],
-} satisfies Config;
-
-export default config;
